fix(ref-imp): fixes controller property to have a value
* fix(ref-imp): fixes controller property to have a value

This is a bit of a hack that's potentially going to create issues when resolving
a key that's not controlled by the controller of the did document. At this point
though that's an exceptional case that doesn't need to be handled immediately and
this fix allows DID Documents resolved by the sidetree node to be compliant with
the did-test-suite.

BREAKING CHANGE: The controller is no longer an empty string and instead matches
the document id property (e.g. the did being resolved)